give tile a separate file
